Separated sparse and dense matrix traits and introduced more dense matrix traits
