+++ data-structures/algebra/src/traits/matrix/dense_matrix.rs
@@ -1,0 +1,6 @@
+//! Submodule defining traits to operate over Dense matrices.
+
+use super::{Matrix, ValuedMatrix};
+
+/// Trait defining a dense matrix.
+pub trait DenseMatrix: Matrix + ValuedMatrix {}
